+++ packages/domain-post/src/test/__test-utils__/mocks/post.repository.mock.ts
@@ -7,8 +7,10 @@ export const createPostRepositoryMock = (): jest.Mocked<IPostRepository> =>
     findByAuthorId: jest.fn(),
     findAll: jest.fn(),
     create: jest.fn(),
+    createWithPostCreated: jest.fn(),
     update: jest.fn(),
     delete: jest.fn(),
+    deleteWithPostDeleted: jest.fn(),
     deleteByAuthorId: jest.fn(),
     search: jest.fn(),
   }) as unknown as jest.Mocked<IPostRepository>;